requirements.txt
absl-py==2.2.2
acres==0.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
comm==0.2.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.0
coremltools==8.3.0
cycler==0.12.1
datasets==3.4.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==4.2.20
duckduckgo_search==8.1.1
etelemetry==0.3.1
evaluate==0.4.3
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.116.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib-metadata==7.0.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.0
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
joblib==1.4.2
jsonpickle==4.1.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.7
llvmlite==0.43.0
looseversion==1.3.0
lxml==5.3.2
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.6
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
nest-asyncio==1.5.9
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.68.2
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==23.2
pandas==2.2.3
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.1.0
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
primp==0.15.0
prompt-toolkit==3.0.43
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.25.6
prov==2.0.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
puremagic==1.28
pyaml==25.7.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.17.2
PyJWT==2.10.1
PyMuPDF==1.25.4
pyparsing==3.2.1
PySocks==1.7.1
pyswip==0.3.2
pytest==8.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.1
pyvis==0.3.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.13.0
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.0
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
simplejson==3.20.1
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
slack_sdk==3.36.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sounddevice==0.5.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.1
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
traits==7.0.2
transformers==4.50.0
twilio==9.6.5
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.17.0
